Move integration tests to unit tests
